safari-600.7-branch/LayoutTests/http/tests/security/contentSecurityPolicy/report-only-report-uri-missing.html

<!DOCTYPE html>
<html>
<head>
    <meta http-equiv="Content-Security-Policy-Report-Only" content="script-src 'unsafe-inline';">
    <script>
        if (window.testRunner)
            testRunner.dumpAsText();
    </script>
</head>
<body>
    <p>This test passes if a console message is present, warning about the missing 'report-uri' directive.</p>
</body>
</html>
